move fuse index building to it's own file
