+++ components/search/Cargo.toml
@@ -8,3 +8,5 @@ errors = { path = "../errors" }
 content = { path = "../content" }
 config = { path = "../config" }
 libs = { path = "../libs" }
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
