+++ components/site/src/lib.rs
@@ -804,61 +804,16 @@ impl Site {
             IndexFormat::ElasticlunrJson | IndexFormat::FuseJson => "json",
         };
         let path = &self.output_path.join(format!("search_index.{}.{}", lang, extension));
+        let library = self.library.read().unwrap();
         let content = match &self.config.search.index_format {
             IndexFormat::ElasticlunrJavascript | IndexFormat::ElasticlunrJson => {
-                search::build_index(lang, &self.library.read().unwrap(), &self.config)?
+                search::build_index(lang, &library, &self.config)?
             }
             IndexFormat::FuseJson | IndexFormat::FuseJavascript => {
-                #[derive(serde::Serialize)]
-                struct Item {
-                    title: Option<String>,
-                    body: Option<String>,
-                    path: String,
-                }
-                let mut items: Vec<Item> = Vec::new();
-                let library = self.library.read().unwrap();
-                let config = &self.config.search;
-                for (_, section) in &library.sections {
-                    if section.lang == lang
-                        && section.meta.redirect_to.is_none()
-                        && section.meta.in_search_index
-                    {
-                        items.push(Item {
-                            path: section.path.clone(),
-                            title: if config.include_title {
-                                Some(section.meta.title.clone().unwrap_or_default())
-                            } else {
-                                None
-                            },
-                            body: if config.include_content {
-                                Some(search::AMMONIA.clean(&section.content).to_string())
-                            } else {
-                                None
-                            },
-                        });
-                        for page in &section.pages {
-                            let page = &library.pages[page];
-                            if page.meta.in_search_index {
-                                items.push(Item {
-                                    title: if config.include_title {
-                                        Some(page.meta.title.clone().unwrap_or_default())
-                                    } else {
-                                        None
-                                    },
-                                    body: if config.include_content {
-                                        Some(search::AMMONIA.clean(&page.content).to_string())
-                                    } else {
-                                        None
-                                    },
-                                    path: page.path.clone(),
-                                })
-                            }
-                        }
-                    }
-                }
-                serde_json::to_string(&items)?
+                search::fuse::build_index(lang, &library, &self.config.search)?
             }
         };
+        drop(library); // no need to hold on to this guard while writing
         create_file(
             path,
             match self.config.search.index_format {
